requirements.txt
requests
beautifulsoup4
readability-lxml
lxml
tiktoken
openai
click
python-dotenv
pytest
